feat(shop): la tienda pasa a su propio apartado del menu del alumno
La tienda de merchandising vivia dentro de la pagina de Retos, que mezclaba
dos cosas distintas: el progreso del alumno y el gasto de sus puntos.

Sale a /shop con entrada propia en el menu lateral, debajo de Retos. Usa
useChallengeSummary en vez de useMyChallenges: solo necesita los puntos, no
la lista completa de retos con su progreso.

En el hero de Retos queda un boton "Canjear puntos" que lleva a la tienda,
para no perder el gesto de ver los puntos y gastarlos. ChallengesPage baja
de ~750 a 341 lineas y se queda con una sola responsabilidad.
